Agrego division entre talonario activos y no activos
